feat: Introduce Router; move App.jsx functionality into Main.jsx
